third_party/WebKit/LayoutTests/fast/events/wheel/wheelevent-ctrl-expected.txt

TOP



Scroll mouse wheel over here



END
And scroll the document here
Tests that wheel events with the ctrl modifier are handled properly

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Test mousewheel events over scrollable div
With ctrl modifier set and canScroll set to be false
PASS wheelEventCount is 1
PASS deltaY is 80
PASS ctrlKey is true
PASS testDiv.scrollTop is 0
Without ctrl and canScroll set to be default true
PASS wheelEventCount is 1
PASS deltaY is 80
PASS ctrlKey is false
PASS testDiv.scrollTop is deltaY

Test mousewheel events over the document
With ctrl modifier set and canScroll set to be false
PASS wheelEventCount is 1
PASS deltaY is 80
PASS ctrlKey is true
PASS window.scrollY is 0
With ctrl modifier set and canScroll set to be true
PASS wheelEventCount is 1
PASS deltaY is 80
PASS ctrlKey is true
PASS window.scrollY is deltaY
Now without ctrl and canScroll set to be default true
PASS wheelEventCount is 1
PASS deltaY is 80
PASS ctrlKey is false
PASS window.scrollY is 160
PASS successfullyParsed is true

TEST COMPLETE
